+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KQCK26Z243W1DJVQ1NQMTX1E"
+        "01KQEDQT8EYMQ4D03ERB9KGJNB"
     )
     r"""The unique identifier of the data source"""