+++ packages/orq-rc/RELEASES.md
@@ -7419,3 +7419,13 @@ Based on:
 - [python v4.9.0-rc.9] packages/orq-rc
 ### Releases
 - [PyPI v4.9.0-rc.9] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.9 - packages/orq-rc
+
+## 2026-04-30 05:30:57
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.9 (2.881.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.9.0-rc.10] packages/orq-rc
+### Releases
+- [PyPI v4.9.0-rc.10] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.10 - packages/orq-rc